chore: Move stuff to routes file
